Cleanup
- removed unneeded "#include Vars.hh"
- moved some code into StressBalance.cc and delete StressBalance_diagnostics.{cc,hh}
