+++ src/coupler/surface/DEBMSimple.cc
@@ -31,7 +31,6 @@
 #include "pism/util/array/CellType.hh"
 #include "pism/util/error_handling.hh"
 #include "pism/util/pism_utilities.hh"
-#include "pism/util/Vars.hh"
 #include "pism/util/array/Forcing.hh"
 #include "pism/util/Logger.hh"
 #include "pism/util/io/IO_Flags.hh"
