+++ src/energy/EnergyModel.cc
@@ -22,7 +22,6 @@
 #include "pism/stressbalance/StressBalance.hh"
 #include "pism/util/MaxTimestep.hh"
 #include "pism/util/Profiling.hh"
-#include "pism/util/Vars.hh"
 #include "pism/util/array/CellType.hh"
 #include "pism/util/error_handling.hh"
 #include "pism/util/io/File.hh"
