+++ src/stressbalance/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_library (stressbalance OBJECT
   StressBalance.cc
-  StressBalance_diagnostics.cc
   ShallowStressBalance.cc
   WeertmanSliding.cc
   SSB_Modifier.cc
